Add tied users to VistA instance creation
Tied users are users who can only get into a VistA environment either as
a normal user (only menu access) which is tied to the ^ZU routine, or as
a programmer (M level access) which is tied on GTM to "mumps -dir". Both
of these options don't allow linux system level access, which other
linux user accounts should be able to accomidate.

Change-Id: I929d4ed9f44e10ab510fe492fdf1b9e174de83fc
OSEHRA-Id: http://issues.osehra.org/browse/VAF-150