Prepare MUnit for intentional failures
Change the ReportUnitTestResult function in the CommonFunctions.cmake file
to account for intentionally failing checks in an MUnit test output.

Add a check to the section that prints the failure line, which attempts to
match the line to the regex "[Ii]ntentional"  If it finds a match, that
error does not cause the test to fail.

Remove  the failure condition that checks for the error number in the
summary.

OSEHRA-Id: http://issues.osehra.org/browse/OAT-150
Change-Id: I5c0bb477111c0a816d87382ab6baeef9bece34d0