Adjust the OTJ Scripts to account for no routines
Adjust the XINDEX,%RCheck, and %RFind function calls to have an expect
statment which allows the code to exit if no routines are found.  This is
best seen when trying to get the before information on a completely new
KIDS patch. The code as-is will throw a timeout error instead.

When the no routines block is accessed, print a message saying that no
routines were found.

OSEHRA-Id: http://issues.osehra.org/browse/OAT-153
Change-Id: I72bb87d421e4c36e4d48fd29856f212f852b3cd5